Changed diff to get closer to finding differences
It will work by asserting equality on the highest level possible.
If there are changes, it will dig down to the items which are different,
drilling down to the scalar item level.

At least that is the plan, not quite sure how that is going to workout
for simple types which are not Json/Any